making albums better and organizing UI

+++ ui/src/app/admin/collection/album-details/album-details.component.ts
@@ -1,20 +1,21 @@
-import { RenditionConfiguration } from './../../models/rendition-configuration';
-import { PhotoService } from './../../services/photo.service';
-import { Photo } from './../../models/photo';
-import { Observable } from 'rxjs/Observable';
-import { Album } from './../../models/album';
-import { Collection } from './../../models/collection';
-import { Subscription } from 'rxjs/Subscription';
-import { CollectionService } from './../../services/collection.service';
-import { AlbumService } from './../../services/album.service';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Component, OnInit, OnDestroy } from '@angular/core';
+
+import { Observable } from 'rxjs/Observable';
+import { Subscription } from 'rxjs/Subscription';
 import 'rxjs/observable/combineLatest';
 import 'rxjs/add/operator/mergeMap';
-import { combineLatest } from 'rxjs/observable/combineLatest';
 import { share } from 'rxjs/operators/share';
+import { combineLatest } from 'rxjs/observable/combineLatest';
+
+import { AlbumStore } from './../../stores/album.store';
 import { Paginator } from '../../models/paginator';
-import { RenditionConfigurationService } from '../../services/rendition-configuration.service';
+import { RenditionConfiguration } from './../../models/rendition-configuration';
+import { Photo } from './../../models/photo';
+import { Album } from './../../models/album';
+import { Collection } from './../../models/collection';
+import { AlbumService } from './../../services/album.service';
+import { CollectionStore } from '../../stores/collection.store';
 
 @Component({
   selector: 'app-album-details',
@@ -25,29 +26,23 @@ export class AlbumDetailsComponent implements OnInit, OnDestroy {
   paginator: Paginator;
 
   album: Observable<Album>;
-  photos: Array<Photo> = [];
+  photos: Observable<Array<Photo>>;
   collection: Collection;
   thumbnailRendition: RenditionConfiguration;
 
+  albumStore: AlbumStore;
+
   constructor(
+    private collectionStore: CollectionStore,
     private route: ActivatedRoute,
     private router: Router,
-    private albumService: AlbumService,
-    private collectionService: CollectionService,
-    private photoService: PhotoService,
-    private renditionConfigService: RenditionConfigurationService
+    private albumService: AlbumService
   ) { }
 
   ngOnInit() {
-
     this.paginator = Paginator.newTimestampPaginator('updated_at');
-    const x = this.collectionService.current.switchMap(collection => {
-      return this.renditionConfigService.forCollection(collection).map(configs => {
-        collection.renditionConfigurations = configs;
-        return collection;
-      });
-    });
-    this.album = combineLatest(x, this.route.params)
+
+    this.album = combineLatest(this.collectionStore.current, this.route.params)
       .switchMap(([collection, params]) => {
         const id = +params['album_id'];
         this.collection = collection;
@@ -56,6 +51,12 @@ export class AlbumDetailsComponent implements OnInit, OnDestroy {
 
         return this.albumService.details(collection, id);
       }).pipe(share());
+
+    this.album.first().subscribe(album => {
+      this.albumStore = this.collectionStore.albumStore(album);
+      this.albumStore.loadPhotos(this.paginator);
+      this.photos = this.albumStore.list;
+    });
   }
 
   ngOnDestroy(): void {
@@ -63,6 +64,9 @@ export class AlbumDetailsComponent implements OnInit, OnDestroy {
   }
 
   delete(album: Album): void {
+    if (!confirm(`Delete album ${album.name}?`)) {
+      return;
+    }
     this.albumService.delete(album.collection, album).subscribe(_ => this.router.navigate([`/admin/collection/${this.collection.slug}`]));
   }
 
@@ -73,5 +77,13 @@ export class AlbumDetailsComponent implements OnInit, OnDestroy {
 
   setCoverPhoto(album: Album, photo: Photo): void {
     this.albumService.setCoverPhoto(album, photo).subscribe(success => console.log('cover photo change success', success));
+  }
+
+  toggleOrganizePhotos(): void {
+    alert('TODO: enter organize mode');
+  }
+
+  shareAlbumDialog(): void {
+    alert('TODO: share album dialog');
   }
 }
